+++ Sources/RAW_md5/MD5.swift
@@ -9,7 +9,7 @@ public struct Hash:Sendable{}
 public struct Hasher<RAW_hasher_outputtype:RAW_staticbuff>:RAW_hasher where RAW_hasher_outputtype.RAW_fixed_type == Hash.RAW_fixed_type {
 	private var context:__crawdog_md5_context
 	
-	public static var RAW_hasher_blocksize:size_t { size_t(__CRAWDOG_MD5_BLOCK_SIZE * 4) }
+	public static var RAW_hasher_blocksize:Int { Int(__CRAWDOG_MD5_BLOCK_SIZE * 4) }
 
 	public typealias RAW_hasher_outputtype = Hash
 
@@ -26,7 +26,7 @@ public struct Hasher<RAW_hasher_outputtype:RAW_staticbuff>:RAW_hasher where RAW_
 		__crawdog_md5_update(&context, buffer.baseAddress!, UInt32(buffer.count))
 	}
 		
-	public mutating func update(_ data:UnsafeRawPointer, count:size_t) {
+	public mutating func update(_ data:UnsafeRawPointer, count:Int) {
 		__crawdog_md5_update(&context, data, UInt32(count))
 	}
