+++ Sources/RAW_sha512/SHA512.swift
@@ -6,7 +6,7 @@ import RAW
 public struct Hash:Sendable {}
 
 public struct Hasher<RAW_hasher_outputtype:RAW_staticbuff>:RAW_hasher where RAW_hasher_outputtype.RAW_fixed_type == (UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8) {
-	public static var RAW_hasher_blocksize:size_t { size_t(__CRAWDOG_SHA512_BLOCK_SIZE) }
+	public static var RAW_hasher_blocksize:Int { Int(__CRAWDOG_SHA512_BLOCK_SIZE) }
 	public typealias RAW_hasher_outputtype = Hash
 
 	private var context:__crawdog_sha512_context
@@ -24,7 +24,7 @@ public struct Hasher<RAW_hasher_outputtype:RAW_staticbuff>:RAW_hasher where RAW_
 		__crawdog_sha512_update(&context, buffer.baseAddress!, UInt32(buffer.count))
 	}
 		
-	public mutating func update(_ data:UnsafeRawPointer, count:size_t) {
+	public mutating func update(_ data:UnsafeRawPointer, count:Int) {
 		__crawdog_sha512_update(&context, data, UInt32(count))
 	}
